Update templates and logic to support gathering and presenting a comment field
This hits a bunch of files:

- lib/view/helpers.js                       <- trimString function for long comments
- views/partials/book_leave_modal.hbs       <- Form with input box for comment
- views/partials/leave_request_detail.hbs   <- Expand the data for insertion into emails
- views/partials/user_requests.hbs          <- Show your own leave calendar
- views/requests.hbs                        <- Manager view

Comments can be arbitrarily long, so we trim the string before display. It's a
bit shorter in the manager interface compared to the user interface.

The comment is also shown in the body of the emails that are sent out. We
considered putting it in the subject line as well, but that could be too ugly
and unwieldy.